feat: Disable dimensions in the build command
The changes in this commit focus on updating the build command in the README file to include the `-disable-dimensions` flag when compiling the `burstingBubble.c` program. This flag is necessary to disable the automatic dimension handling in the compiler. I prefer working with dimensionless variables and usually know what I am doing.. :D